custom-elements/registering-custom-elements/definition-construction-algorithm-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>Default namespace is HTML namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Default namespace is HTML namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var XFoo = doc.register('X-FOO-1');

    var xfoo = new XFoo();

    assert_equals(xfoo.namespaceURI, HTML_NAMESPACE, 'Wrong namespace');

}, 'Namespace should be HTML namespace');

</script>
</body>
</html>
